+++ src/lib/web3/utils/ticks.ts
@@ -2,16 +2,46 @@ import BigNumber from 'bignumber.js';
 import { Token } from './tokens';
 
 /**
- * TickMap contains a mapping from tickIDs to tick indexes inside poolsZeroToOne and poolsOneToZero
+ * price1To0:
+ *
+ * we center our definition of price1To0 from the chain definition
+ * in link: https://github.com/duality-labs/duality/blob/v0.1.5/x/dex/keeper/core_helper.go#L148-L151
+ * price1To0 helps use convert token1 amounts to equivalent token0 amounts
+ * for relative value calculations, hence the name "1 to 0"
+ * $ totalAmount0 = amount0 + amount1 * price1To0
+ *
+ * we can explain further with an example across a few orders of magnitude:
+ *  - take token0 as ATOM
+ *  - take token1 as ETH
+ *  - take price0 = 10USD/ATOM
+ *  - take price1 = 1000USD/ETH
+ *
+ * $ totalAmount0 = amount0 + amount1 * price1To0
+ * $ totalAmountATOM = amountATOM + amountETH * priceATOMToETH
+ * $ totalAmountATOM = amountATOM + amountETH * (priceUSD/ETH / priceUSD/ATOM)
+ * $ totalAmountATOM = amountATOM + amountETH * (price1 / price0)ATOM/ETH
+ * $ totalAmountATOM = amountATOM + amountETH * 100ATOM/ETH
+ *
+ * therefore: price1To0 = price1 / price0
  */
-export interface TickMap {
-  [tickID: string]: PoolTicks;
-}
+type Price = BigNumber;
 
-type PoolTicks = [
-  index0to1: TickInfo | undefined,
-  index1to0: TickInfo | undefined
-];
+/**
+ * tickIndex1To0:
+ *
+ * tickIndex1To0 represents the tickIndex of the tokenPair 0<>1 at price1To0
+ * price = 1.0001^[tickIndex] and
+ * tickIndex = log(price) / log(1.0001)
+ *  # for direction 1To0 that means:
+ * price1To0 = 1.0001^[tickIndex1To0] and
+ * tickIndex1To0 = log(price1To0) / log(1.0001)
+ *  # from our example token pair ATOM<>ETH:
+ * tickIndex1To0 = log(200) / log(1.0001)
+ * tickIndex1To0 = 46054 (always rounded to an integer)
+
+ * importantly the API will usually return tickIndex1To0 as tickIndex
+ */
+type TickIndex = BigNumber;
 
 /**
  * TickInfo is a reflection of the backend structue "DexPool"
@@ -23,8 +53,13 @@ export interface TickInfo {
   reserve0: BigNumber;
   reserve1: BigNumber;
   fee: BigNumber;
-  tickIndex: BigNumber; // tickIndex is the exact price ratio in the form: 1.0001^[tickIndex]
-  price: BigNumber; // price is an approximate decimal (to 18 places) ratio of price1/price0
+  // price1To0 is an approximate decimal (to 18 places) ratio of price1/price0
+  // see Price definition above
+  price1To0: Price;
+  // tickIndex1To0 is an integer representing price1To0 using the equation:
+  //   price = 1.0001^[tickIndex]
+  // see TickIndex definition above
+  tickIndex1To0: TickIndex;
 }
 
 export function tickIndexToPrice(tickIndex: BigNumber): BigNumber {
@@ -44,21 +79,21 @@ export function priceToTickIndex(
 
 const bigZero = new BigNumber(0);
 export function calculateShares({
-  price,
-  tickIndex,
+  price1To0,
+  tickIndex1To0,
   reserve0 = bigZero,
   reserve1 = bigZero,
 }: Partial<TickInfo> &
-  // must include either price or tickIndex
+  // must include either price1To0 or tickIndex1To0
   (| {
-        price: TickInfo['price'];
+        price1To0: TickInfo['price1To0'];
       }
     | {
-        price?: undefined;
-        tickIndex: TickInfo['tickIndex'];
+        price1To0?: undefined;
+        tickIndex1To0: TickInfo['tickIndex1To0'];
       }
   )): BigNumber {
   return reserve0.plus(
-    reserve1.multipliedBy(price || tickIndexToPrice(tickIndex))
+    reserve1.multipliedBy(price1To0 || tickIndexToPrice(tickIndex1To0))
   );
 }